[droidian-meta] Add a workaround to dynamically pull Qt's Wayland QPA whenever a Qt application gets installed
Droidian is currently a mostly-GTK+ distro, so we try to avoid
foreign dependencies if not strictly needed.

This has the side effect that, when Qt applications are installed, no
working backend is put in place (since applications don't depend on any),
resulting in a worsened end-user experience.

This commit adds a workaround that bases itself on the inner APT workings:
droidian-minimal now depends on `droidian-no-qt | droidian-qt`, with the
former getting pulled up automatically during image creation, and that
declares a conflict relationship with the core Qt libraries.

When an end-user is going to install a Qt application, APT will resolve
the conflict by installing the alternative `droidian-qt` package instead,
which will pull qtwayland5 as well.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>