[droidian-phosh-minimal] (THIS IS NOT NEEDED) Depend on adaptation-hybris-phosh
Recent changes to adaptation-hybris split the old adaptation-hybris metapackage
to -common and -phosh. There is an upgrade path, and during installation -phosh
is automatically marked as manually installed (as it should be in
new images).

Some esoteric APT alternatives (like aptitude) might want to drop the -phosh
package during the initial install. This is a theorical issue, and this dependency
has been put here to be in the safe side. This can be reverted safely in a couple
of weeks.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>